+++ examples/data-sources/iosxe_crypto_engine/data-source.tf
@@ -1,2 +1,0 @@
-data "iosxe_crypto_engine" "example" {
-}