+++ examples/resources/iosxe_crypto_engine/import.sh
@@ -1,1 +1,0 @@
-terraform import iosxe_crypto_engine.example ""