+++ examples/resources/iosxe_crypto_engine/resource.tf
@@ -1,3 +1,0 @@
-resource "iosxe_crypto_engine" "example" {
-  compliance_shield_disable = true
-}