+++ docs/guides/changelog.md
@@ -7,9 +7,13 @@ description: |-
 
 # Changelog
 
+## (Unreleased)
+
+- Add `ip_local_proxy_arp` attribute to `iosxe_interface_vlan` resource and data sources
+- Add `iosxe_crypto` resource and data source
+
 ## 0.11.0
 
-- Add `iosxe_crypto_engine` resource and data source for crypto engine compliance shield disable configuration
 - Fix `iosxe_yang` resource payload ordering with NETCONF, [link](https://github.com/CiscoDevNet/terraform-provider-iosxe/issues/372)
 - Add `ip_igmp_version` attribute to `iosxe_interface_ethernet`, `iosxe_interface_loopback`, `iosxe_interface_port_channel`, `iosxe_interface_port_channel_subinterface`, `iosxe_interface_tunnel`, and `iosxe_interface_vlan` resources and data sources
 - Add `ip_default_gateway` attribute to `iosxe_system` resource and data source for default gateway configuration on non-routing devices